+++ Assets/Scripts/Safe/ScoreManager.cs
@@ -7,20 +7,20 @@ namespace Safe
     {
         private const string saveFileName = "player_score.txt";
 
-        public static void SaveScore(int score)
+        public static void SaveScore(ulong score)
         {
             string saveFilePath = Path.Combine(Application.persistentDataPath, saveFileName);
             File.WriteAllText(saveFilePath, score.ToString());
         }
 
-        public static int LoadScore()
+        public static ulong LoadScore()
         {
             string saveFilePath = Path.Combine(Application.persistentDataPath, saveFileName);
             if (File.Exists(saveFilePath))
             {
                 string savedScore = File.ReadAllText(saveFilePath);
-                int result;
-                if (int.TryParse(savedScore, out result))
+                ulong result;
+                if (ulong.TryParse(savedScore, out result))
                 {
                     return result;
                 }
